+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### unreleased
+### Added
+* Added an option to set the compression quality for all files.
+
 ### Fixed
 * Fixed an issue where some pixels near the edges would be incorrectly set as transparent when decoding or changing formats.